+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# Next
+- Added a command to change categories (associate new to old when some change is decided).
+
 # 2.0.0 (2020-03-30)
 - Added maps! (by @desjarlaisdumaishugo)
 - Added an artisan command to loop through database to geocode address.